Add support for broadcasting without an active session
