+++ Realtime/src/commonMain/kotlin/io/github/jan/supabase/realtime/data/PostgresActionData.kt
@@ -1,5 +1,6 @@
-package io.github.jan.supabase.realtime
+package io.github.jan.supabase.realtime.data
 
+import io.github.jan.supabase.realtime.Column
 import kotlinx.datetime.Instant
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
